requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchvision==0.21.0
accelerate==0.20.2
diffusers==0.20.0
transformers==4.30.2
huggingface_hub==0.14.*
datasets
tensorboard
xformers
hpsv2
image-reward
open-clip-torch==2.16.0
torchmetrics
openai_clip==1.0.1